Removed Braille folder and refactored code, merging it to WatchApp folder
