--- Braille/.idea/misc.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="DesignSurface">
-    <option name="filePathToZoomLevelMap">
-      <map>
-        <entry key="app/src/main/res/layout/activity_main.xml" value="0.2961515154157366" />
-        <entry key="app/src/main/res/layout/activity_read.xml" value="0.19874476987447698" />
-        <entry key="app/src/main/res/layout/activity_write.xml" value="0.5324267782426778" />
-        <entry key="app/src/main/res/layout/fragment_main.xml" value="0.5324267782426778" />
-        <entry key="app/src/main/res/layout/n.xml" value="0.19874476987447698" />
-      </map>
-    </option>
-  </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
-    <output url="file://$PROJECT_DIR$/build/classes" />
-  </component>
-  <component name="ProjectType">
-    <option name="id" value="Android" />
-  </component>
-</project>
