--- Braille/.idea/vcs.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="VcsDirectoryMappings">
-    <mapping directory="$PROJECT_DIR$/UCSD-CSE-118-218-Team-G" vcs="Git" />
-  </component>
-</project>
