--- Braille/app/src/main/res/values/themes.xml
@@ -1,6 +1,0 @@
-<resources>
-
-    <style name="AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
-
-    <style name="PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
-</resources>
