--- Braille/build.gradle
@@ -1,9 +1,0 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-plugins {
-    id 'com.android.application' version '7.2.0' apply false
-    id 'com.android.library' version '7.2.0' apply false
-}
-
-task clean(type: Delete) {
-    delete rootProject.buildDir
-}
