--- Braille/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Mon Nov 20 11:28:40 PST 2023
-distributionBase=GRADLE_USER_HOME
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.3.3-bin.zip
-distributionPath=wrapper/dists
-zipStorePath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
