+++ WatchApp/app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
 
     <uses-feature android:name="android.hardware.type.watch" />
     <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.VIBRATE" />
 
     <application
         android:allowBackup="true"
@@ -33,6 +34,15 @@
             android:exported="true">
         </activity>
         <activity android:name=".UserListActivity"
+            android:exported="true">
+        </activity>
+        <activity android:name=".BrailleOption"
+            android:exported="true">
+        </activity>
+        <activity android:name=".BrailleMessageReceive"
+            android:exported="true">
+        </activity>
+        <activity android:name=".BrailleMessageSend"
             android:exported="true">
         </activity>
     </application>
