+++ WatchApp/build.gradle
@@ -1,4 +1,4 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-id 'com.android.application' version '8.1.2' apply false
+id 'com.android.application' version '8.1.4' apply false
 }
